--- src/Components/test.js
@@ -1,74 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { useNavigate } from 'react-router-dom';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import CardMedia from '@mui/material/CardMedia';
-import Typography from '@mui/material/Typography';
-import { Button, CardActionArea, CardActions } from '@mui/material';
-import { Icon } from '@iconify/react';
-// import BeeItemSocial from './BeeItemSocial';
-
-const BeeItem = (props) => {
-  const { bee } = props;
-  const navigate = useNavigate();
-
-  const redirect = () => {
-    navigate(
-      `/${bee.id}`,
-      { state: { bee } },
-    );
-  };
-
-  const imageStyle = {
-    height: 0,
-    paddingTop: '80%',
-    width: '100%',
-    objectFit: 'cover',
-  };
-
-  const cardStyle = {
-    margin: '0 20px',
-    height: '100%',
-    display: 'flex',
-    justifyContent: 'space-between',
-    flexDirection: 'column',
-    alignItems: 'center',
-  };
-
-  return (
-    <Card style={cardStyle}>
-      <CardActionArea onClick={redirect}>
-        <CardMedia image={bee.image} title={bee.name} style={imageStyle} />
-        <CardContent>
-          <Typography gutterBottom variant="h5" component="h2">
-            {bee.name}
-          </Typography>
-          <Typography component="p">{bee.description}</Typography>
-        </CardContent>
-      </CardActionArea>
-      <CardActions>
-        <Button size="large" color="primary">
-          <Icon icon="bi:facebook" />
-        </Button>
-        <Button size="large" color="primary">
-          <Icon icon="bi:twitter" />
-        </Button>
-        <Button size="large" color="primary">
-          <Icon icon="bi:instagram" />
-        </Button>
-      </CardActions>
-    </Card>
-  );
-};
-
-BeeItem.propTypes = {
-  bee: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    description: PropTypes.string.isRequired,
-    image: PropTypes.string.isRequired,
-  }).isRequired,
-};
-
-export default BeeItem;
